Added Jak 1 autosplit options for completing specific tasks that are turned in at a hub level (#22)
Added Jak 1 autosplit options for completing specific tasks that are
turned in at a hub level, such as catching the muse, redirecting the eco
beams in Jungle, and completing the moles and gambler's race tasks in
Basin. I moved the splits for turning in these tasks to the hub levels
where you actually get the power cells.